requirements.txt
fastapi
uvicorn
sqlalchemy

pyjwt

jinja2

python-multipart